move test to different file
